requirements.txt
pillow
openai